Fix tests for jdk7
